Correcting typo in definition of msg_type macro in Messages.h.  Updating unit test accordingly to accommodate.

+++ lib/cnext/include/Messages.h
@@ -38,6 +38,7 @@
 
 #include "stdint.h"
 #include "stdbool.h"
+#include "stdlib.h"
 #include "time.h"
 
 // Do forward declarations before including headers from this library to avoid
@@ -236,7 +237,7 @@ msg_t* msg_wait_for_reply_with_type(msg_t *sent,
 // Message element accessors
 void* msg_element(msg_t *msg, msg_element_t msg_element);
 #define msg_type(msg_ptr) \
-  (*((int*) msg_element((msg_ptr), MSG_ELEMENT_TYPE)))
+  (*((int64_t*) msg_element((msg_ptr), MSG_ELEMENT_TYPE)))
 #define msg_data(msg_ptr) \
   (*((void**) msg_element((msg_ptr), MSG_ELEMENT_DATA)))
 #define msg_size(msg_ptr) \
